roles: extract RoleFormFields for new + edit pages
