+++ src/routes/(app)/database/character-roles/[id]/edit/+page.svelte
@@ -20,32 +20,13 @@
 	import ModalBody from '$lib/components/ui/ModalBody.svelte'
 	import ModalFooter from '$lib/components/ui/ModalFooter.svelte'
 	import DatabaseFormHeader from '$lib/components/database/DatabaseFormHeader.svelte'
-	import RoleIcon from '$lib/components/database/RoleIcon.svelte'
-	import DetailsContainer from '$lib/components/ui/DetailsContainer.svelte'
-	import DetailItem from '$lib/components/ui/DetailItem.svelte'
+	import RoleFormFields from '$lib/components/database/RoleFormFields.svelte'
 
 	import { localizeHref } from '$lib/paraglide/runtime'
 	import { extractErrorMessage } from '$lib/utils/errors'
-	import {
-		dataUrlToBase64,
-		validateIconFile,
-		type IconValidationError
-	} from '$lib/utils/iconUpload'
+	import { dataUrlToBase64 } from '$lib/utils/iconUpload'
 
 	import type { PageData } from './$types'
-
-	function iconErrorMessage(error: IconValidationError): string {
-		switch (error) {
-			case 'mime':
-				return m.roles_icon_error_png()
-			case 'size':
-				return m.roles_icon_error_size()
-			case 'dimensions':
-				return m.roles_icon_error_dimensions()
-			case 'decode':
-				return m.roles_icon_error_png()
-		}
-	}
 
 	let { data }: { data: PageData } = $props()
 
@@ -67,11 +48,7 @@
 	let iconFile = $state<File | null>(null)
 	let iconPreview = $state<string | null>(null)
 	let iconError = $state<string | null>(null)
-	let iconInputRef = $state<HTMLInputElement | undefined>(undefined)
 
-	function openIconPicker() {
-		iconInputRef?.click()
-	}
 	let isSaving = $state(false)
 	let saveError = $state<string | null>(null)
 	let saveSuccess = $state(false)
@@ -92,21 +69,10 @@
 		}
 	})
 
-	async function handleIconSelect(e: Event) {
+	function handleSelectIcon(selection: { file: File; dataUrl: string }) {
 		iconError = null
-		const input = e.target as HTMLInputElement
-		const file = input.files?.[0]
-		if (!file) return
-
-		const result = await validateIconFile(file)
-		if (!result.ok) {
-			iconError = iconErrorMessage(result.error)
-			input.value = ''
-			return
-		}
-
-		iconFile = result.file
-		iconPreview = result.dataUrl
+		iconFile = selection.file
+		iconPreview = selection.dataUrl
 	}
 
 	async function handleSave() {
@@ -188,61 +154,14 @@
 		{/if}
 
 		<section class="details">
-			<DetailsContainer title={m.roles_section_basics()}>
-				<DetailItem
-					label={m.roles_field_name_en()}
-					bind:value={editData.nameEn}
-					editable={true}
-					type="text"
-					placeholder={m.roles_placeholder_name_en()}
-				/>
-				<DetailItem
-					label={m.roles_field_name_jp()}
-					bind:value={editData.nameJp}
-					editable={true}
-					type="text"
-					placeholder={m.roles_placeholder_name_jp()}
-				/>
-			</DetailsContainer>
-
-			<DetailsContainer title={m.roles_section_icon()}>
-				<div class="icon-upload">
-					<button
-						type="button"
-						class="icon-trigger"
-						onclick={openIconPicker}
-						aria-label={m.roles_icon_upload()}
-					>
-						<RoleIcon
-							iconKey={role.iconKey}
-							src={iconPreview ?? undefined}
-							name={role.nameEn}
-							size={64}
-							imageSize={48}
-						/>
-					</button>
-
-					<input
-						bind:this={iconInputRef}
-						type="file"
-						accept="image/png"
-						onchange={handleIconSelect}
-						class="visually-hidden"
-						aria-hidden="true"
-						tabindex="-1"
-					/>
-
-					<div class="icon-controls">
-						<Button variant="secondary" size="small" onclick={openIconPicker}>
-							{m.roles_icon_choose()}
-						</Button>
-						<p class="hint">{m.roles_icon_hint_edit()}</p>
-						{#if iconError}
-							<p class="icon-error">{iconError}</p>
-						{/if}
-					</div>
-				</div>
-			</DetailsContainer>
+			<RoleFormFields
+				bind:values={editData}
+				existingIconKey={role.iconKey}
+				{iconPreview}
+				{iconError}
+				onSelectIcon={handleSelectIcon}
+				onIconError={(msg) => (iconError = msg)}
+			/>
 
 			<div class="delete-row">
 				<Button variant="destructive" size="small" onclick={() => (confirmDeleteOpen = true)}>
@@ -277,7 +196,6 @@
 <style lang="scss">
 	@use '$src/themes/spacing' as spacing;
 	@use '$src/themes/layout' as layout;
-	@use '$src/themes/typography' as typography;
 	@use '$src/themes/database' as database;
 
 	.page {
@@ -296,56 +214,6 @@
 
 	.success-banner {
 		@include database.success-banner;
-	}
-
-	.icon-upload {
-		display: flex;
-		gap: spacing.$unit-2x;
-		align-items: center;
-		padding: spacing.$unit-2x;
-	}
-
-	.icon-trigger {
-		padding: 0;
-		border: none;
-		background: transparent;
-		cursor: pointer;
-		border-radius: layout.$item-corner;
-
-		&:focus-visible {
-			outline: 2px solid var(--accent-blue);
-			outline-offset: 2px;
-		}
-	}
-
-	.visually-hidden {
-		position: absolute;
-		width: 1px;
-		height: 1px;
-		padding: 0;
-		margin: -1px;
-		overflow: hidden;
-		clip: rect(0, 0, 0, 0);
-		white-space: nowrap;
-		border: 0;
-	}
-
-	.icon-controls {
-		display: flex;
-		flex-direction: column;
-		gap: spacing.$unit;
-
-		.hint {
-			margin: 0;
-			font-size: typography.$font-small;
-			color: var(--text-tertiary);
-		}
-
-		.icon-error {
-			margin: 0;
-			font-size: typography.$font-small;
-			color: var(--error, #e53e3e);
-		}
 	}
 
 	.delete-row {
